pyproject.toml
[project]
name = "llm-api-switch"
version = "0.1.0"
description = "Free LLM API Auto-Aggregator — auto-discover, monitor, and route free LLM APIs"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "litellm>=1.40.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "pyyaml>=6.0",
    "markdown-it-py>=3.0.0",
    "apscheduler>=3.10.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
